+++ src/pages/BookmarkPage.js
@@ -1,5 +1,13 @@
+import { useSelector } from "react-redux";
+import SubPageTemplate from "../components/SubPageTemplate";
+
 function BookmarkPage() {
-  return <></>;
+  const bookmarkList = useSelector((state) => state.bookmarkList);
+  return (
+    <>
+      <SubPageTemplate baseList={bookmarkList} />
+    </>
+  );
 }
 
 export default BookmarkPage;
